Test: 대분류 목록 조회

+++ back/src/main/java/com/gamul/api/service/IngredientService.java
@@ -10,7 +10,7 @@ import com.gamul.db.entity.HighClass;
 import java.util.List;
 
 public interface IngredientService {
-    List<IngredientInfoRes> getIngredientList(int orderType, int highClassId);
+    List<IngredientInfoRes> getIngredientList(int orderType, Long highClassId);
     List<IngredientInfoRes> getIngredientSelectedList(String userName);
 
     IngredientDetailRes getIngredientDetailInfo(Long ingredientId);
